refactor(api): centralize SDK type overrides into sdk-overrides
